[dt,clkmgr] Move rstmgr dt binding to hw/top/dt instead of ipgening it
